vSomeIP 2.0.0

+++ implementation/utility/include/byteorder.hpp
@@ -17,37 +17,37 @@
 
 #if __BYTE_ORDER == __LITTLE_ENDIAN
 
-#define VSOMEIP_BYTES_TO_WORD(x0, x1) ((x0) << 8 | (x1))
-#define VSOMEIP_BYTES_TO_LONG(x0, x1, x2, x3) ((x0) << 24 | (x1) << 16 | (x2) << 8 | (x3))
+#define VSOMEIP_BYTES_TO_WORD(x0, x1) (uint16_t((x0) << 8 | (x1)))
+#define VSOMEIP_BYTES_TO_LONG(x0, x1, x2, x3) (uint32_t((x0) << 24 | (x1) << 16 | (x2) << 8 | (x3)))
 
-#define VSOMEIP_WORDS_TO_LONG(x0, x1) ((x0) << 16 | (x1))
+#define VSOMEIP_WORDS_TO_LONG(x0, x1) (uint32_t((x0) << 16 | (x1)))
 
-#define VSOMEIP_WORD_BYTE0(x) ((x) & 0xFF)
-#define VSOMEIP_WORD_BYTE1(x) ((x) >> 8)
+#define VSOMEIP_WORD_BYTE0(x) (uint8_t((x) & 0xFF))
+#define VSOMEIP_WORD_BYTE1(x) (uint8_t((x) >> 8))
 
-#define VSOMEIP_LONG_BYTE0(x) ((x) & 0xFF)
-#define VSOMEIP_LONG_BYTE1(x) (((x) >> 8) & 0xFF)
-#define VSOMEIP_LONG_BYTE2(x) (((x) >> 16) & 0xFF)
-#define VSOMEIP_LONG_BYTE3(x) (((x) >> 24) & 0xFF)
+#define VSOMEIP_LONG_BYTE0(x) (uint8_t((x) & 0xFF))
+#define VSOMEIP_LONG_BYTE1(x) (uint8_t(((x) >> 8) & 0xFF))
+#define VSOMEIP_LONG_BYTE2(x) (uint8_t(((x) >> 16) & 0xFF))
+#define VSOMEIP_LONG_BYTE3(x) (uint8_t(((x) >> 24) & 0xFF))
 
-#define VSOMEIP_LONG_WORD0(x) ((x) & 0xFFFF)
-#define VSOMEIP_LONG_WORD1(x) (((x) >> 16) & 0xFFFF)
+#define VSOMEIP_LONG_WORD0(x) (uint16_t((x) & 0xFFFF))
+#define VSOMEIP_LONG_WORD1(x) (uint16_t(((x) >> 16) & 0xFFFF))
 
 #elif __BYTE_ORDER == __BIG_ENDIAN
 
-#define VSOMEIP_BYTES_TO_WORD(x0, x1) ((x1) << 8 | (x0))
-#define VSOMEIP_BYTES_TO_LONG(x0, x1, x2, x3) ((x3) << 24 | (x2) << 16 | (x1) << 8 | (x0))
+#define VSOMEIP_BYTES_TO_WORD(x0, x1) (uint16_t((x1) << 8 | (x0)))
+#define VSOMEIP_BYTES_TO_LONG(x0, x1, x2, x3) (uint32_t((x3) << 24 | (x2) << 16 | (x1) << 8 | (x0)))
 
-#define VSOMEIP_WORD_BYTE0(x) ((x) >> 8)
-#define VSOMEIP_WORD_BYTE1(x) ((x) & 0xFF)
+#define VSOMEIP_WORD_BYTE0(x) (uint8_t((x) >> 8))
+#define VSOMEIP_WORD_BYTE1(x) (uint8_t((x) & 0xFF))
 
-#define VSOMEIP_LONG_BYTE0(x) (((x) >> 24) & 0xFF)
-#define VSOMEIP_LONG_BYTE1(x) (((x) >> 16) & 0xFF)
-#define VSOMEIP_LONG_BYTE2(x) (((x) >> 8) & 0xFF)
-#define VSOMEIP_LONG_BYTE3(x) ((x) & 0xFF)
+#define VSOMEIP_LONG_BYTE0(x) (uint8_t(((x) >> 24) & 0xFF))
+#define VSOMEIP_LONG_BYTE1(x) (uint8_t(((x) >> 16) & 0xFF))
+#define VSOMEIP_LONG_BYTE2(x) (uint8_t(((x) >> 8) & 0xFF))
+#define VSOMEIP_LONG_BYTE3(x) (uint8_t((x) & 0xFF))
 
-#define VSOMEIP_LONG_WORD0(x) ((((x) >> 16) & 0xFFFF)
-#define VSOMEIP_LONG_WORD1(x) (((x) & 0xFFFF)
+#define VSOMEIP_LONG_WORD0(x) (uint16_t((((x) >> 16) & 0xFFFF))
+#define VSOMEIP_LONG_WORD1(x) (uint16_t(((x) & 0xFFFF))
 
 #else
